WIP: add clang/mingw build path, fix latent x64 and K&R portability bugs
  Reconstructs the two codegen scripts that were never committed (blocking
  every compiler, MSVC included) and adds x64-clang-mingw presets.

  Does not build yet; still iterating on remaining errors.
  Note: the sdmparse.h prototype fix changes MSVC behaviour too.

+++ src/Opus/etcmd.c
@@ -218,8 +218,8 @@ CMB * pcmb;
 		{
 		/* initialize the cab */
 		pcabThesaurus = (CABTHESAURUS *) *pcmb->hcab;
-		(int)(pcabThesaurus->uSynList) = LB_ERR;
-		(int)(pcabThesaurus->uDefList) = LB_ERR;
+		pcabThesaurus->uSynList = LB_ERR;
+		pcabThesaurus->uDefList = LB_ERR;
 		if (!FSetCabSz(pcmb->hcab, szEmpty, Iag(CABTHESAURUS, hszLookup)))
 			return cmdNoMemory;
 		}
